requirements.txt
git+https://github.com/potatolondon/djangae.git#egg=djangae
django>=1.7,<1.9
django-secure
git+https://github.com/mozilla/django-csp.git@f728d3af0b94a75ea9fb0bc88efbd74a45726000#egg=django-csp
git+https://github.com/adamalton/django-csp-reports.git#egg=cspreports
django-session-csrf
nose
six
djangorestframework
django-filter
factory-boy
django-webtest
mock
requests
requests-toolbelt
